refactor(eval): split clado backend
